Small-scale polishes and cleanups
Feel free to use any or all of these changes in v0.4 if you don't want to touch v0.3. I didn't change any behavior, but updated some cosmetics:

 - the print message now shows the (i) option.
 - the input carets now have a space before user input starts
 - some Python idioms (`while True`, `if some_string:`, generator comprehensions, etc)
 - the command for writing an image now says where the file was created
 - some string concatenation avoided by using `str.join`

The last bullet should be an efficiency improvement. It's much faster to accumulate strings in a list and then join them all at once than to concatenate each one in a loop.